fix(cost centers): apply E2E renames
Fixes a bug with references in the end to end test names.
